Pass to align transfer_reads (#867)
This PR adds the necessary pass to align `transfer_read`. It is based on
mlir-aie's aievec, see:


https://github.com/Xilinx/mlir-aie/blob/d3da586305ebc22e5ecdf1d3e682b44853436e91/lib/Dialect/AIEVec/Transforms/VectorToVectorConversions.cpp#L123

Some changes were needed for our use case, however. The main one is that
the lowering in this PR skips the `vector.extract_strided_slice`
operation, because we have an offset which is not constant. i.e. the
offsets in
https://mlir.llvm.org/docs/Dialects/Vector/#vectorextract_strided_slice-vectorextractstridedsliceop
cannot be integers for us, because they are determined from loop
induction variables. The pass implemented here goes straight to aievec
extract and shift operations, where mlir Values are used for offsets.

Also included in this PR: an aievec.shift folder. I can make this a
separate PR if preferred.

This PR enables vectorization for convolution and resolves
https://github.com/nod-ai/iree-amd-aie/issues/820