Split L2 input and output objectFifos for memTile/shimTile distribution (#903)
This is the first PR needed to enable 4x4 AIE array. The L2 objectFifos
are split to distribute on multiple memTiles and shimTiles for more
channel usage. The shim/mem tile reassignment will be addressed in a
separate PR.

Also note this PR doesn't change/combine the previous pass that
splitting the third input (elementwise) for connection reuse with mamtul
ops. What unclear to me is how matmul-elementwise path will be like for
4x4 AIE cores. If the existing logic is kept, we'll have to split the
elementwise op to 16 objectFifos which will be a big challenge given the
number of shimTile channels . I'd rather leave it as a separate thing
for now before we have a clear path to move forward.

---------

Co-authored-by: yzhang93 <vivizhan@amd.com>