[AMDAIEAccessToAcquireRelease] Insert acquire/release pairs in same scope (#904)
Before this PR, the logic for read access was invalid. Code like 

```
for i in range(10):
    read_access
read_access
```

got transformed to 

```
for i in range(10):
    acquire
    read_access
release
acquire
read_access
release
```

As you can see above, the code ignored the scoping of the accesses. This
PR ensures that acquires and releases are inserted in the same scope.

---------

Co-authored-by: Jorn Tuyls <jtuyls@users.noreply.github.com>