Add pass to assign tiles to logical objectFifos (#915)
Adds a pass that assigns tiles to logical objectFifos based on tile
usage by trying to to distribute logical objectFifos as evenly as
possible across available tiles. The new pass is based on the tile
assignment logic that was originally part of
`DistributeCoresAndObjectFifos` and refactors the latter pass to use the
extracted functions to keep the logic the same for now.

Note that I am not completely removing the tile assignment functions
from `DistributeCoresAndObjectFifos` for two reasons:
- To keep the tests largely the same for now.
- Because the output of `DistributeCoresAndObjectFifos` before tile
assignment is not stable, i.e. calling cse/canonicalize in between
`DistributeCoresAndObjectFifos` and `AssignTiles` leads to different
functional (incorrect) results.

Until the output of `DistributeCoresAndObjectFifos` before tile
assignment is stable, the tile assignment utility functions should be
called within the pass itself and not be separated out.