[CI] Print program memory usage (PM) along with kernel time (#953)
This will be useful to see how close we are to overflowing program
memory (16384 B). In the case where we do overflow, we won't see how
much PM we require. For that, we'll need to add a check in aie2xclbin --
future work.

UPDATE: the size if the .elf file isn't what we want, the elf file
contains other stuff. I discovered that the program memory size is
stored at byte 72 of the elf (see
https://github.com/newling/aie-rt/commit/d0f08bc4a37092a919d6a0d51a44d9f0ae274bb9
for how I found this)