[AMDAIELinalgFunctionOutlining] Don't outline if operand has non-identity layout  (#959)
This is a fix for a bug I introduced in
https://github.com/nod-ai/iree-amd-aie/pull/951 . I thought that the
approach of removing the offset from the memref function signature
worked, but it doesn't seem to. I must have been running a test other
than the convolution end-to-end test when I was confirming it worked -
now with function outlining enabled in gives a numerical error.

This PR simplifies logic: any non-identity layout results in no
outlining.