[AMDAIEFusePackIntoLoop] Handle case where there is no extract_slice  (#983)
This is very similar to https://github.com/nod-ai/iree-amd-aie/pull/976 

When the AIE grid we're using is `m x n` where one or both of `m` and
`n` is `1`, for a matmul we get pack ops that produce operands for
matmuls directly. i.e. as opposed to `pack->extract_slice->matmul` we
have `pack->matmul`.