[KernelDispatch] Refactor the pass for generic ops and utils for matmul_transpose_b ops (#991)
This PR refactored and simplified the root config setting in
`KernelDispatch.cpp` for the following aspects:

- Removed `setTransposeLikeOpRootConfig` and flag `isMatmulTranspseB`.
Instead, directly go through `setRootConfig` for generic version as well
as the named op version of `matmul_transpse_b`. This will be easier when
we extend the support to cover `matmul_transpose_a`.
- Moved the `isMatmulTransposeB` utils to `Utils.cpp` and reuse/merge
with the existing utils for matmul ops.
- Added missing tests for generic version of `matmul` and
`matmul_transpose_b`.