Add NpuTctSync operation (#990)
This PR introduces the `tct_sync` operation, designed to replace
`dma_wait` during the `ControlCodeLowering` pass. The primary motivation
for this change is to ensure a one-to-one conversion in the subsequent
`ControlCodeToTransaction` pass.

Checks for the same row, channel, and direction are now handled during
the `ControlCodeLowering` pass instead of the `ControlCodeToTransaction`
pass. Some unit tests have also been relocated.