[CombineStridedOps] Generalize dimension checking (#1032)
This PR generalizes dimension checking in `CombineStridedOps`. This
allows different ops to provide their own dimension checking
implementation. This is needed because circular and non-circular
`dma_cpy_nd` ops have different logic for checking whether the number of
dimensions of the combined DMA op would exceed the available number of
dimensions.

With this fix, BatchMatmul works with the `pack-peel-4-level-tiling`
pipeline.