[CI] Disable turbo mode tests due to instability (#1049)
Locally, on a Strix machine, I have been able to crash the NPU in turbo
mode, requiring a cold reboot. I am not sure that this is the same issue
as we're occasionally seeing with the CI runners, but the result is
similar, and I think it's worth a try disabling these tests for a while
to verify.