Initial support for using peano with AIE2P (strix).  (#1071)
The majority of this PR is refactoring to make it possible to support 2+
target ISAs.

- aievec.matmul now verifies its operand shapes based on the target
device in the module.
- I've removed a bunch of intrinsic matmul shapes for phoenix that we're
not using, to simplify the code
- The XLLVM dialect ops now include the target device in their names.
i.e. the name includes either `AIE2` or `AIE2P` now.
- Only matmul can lower to XLLVM from aievec for AIE2P, all other aievec
ops (UPS, etc) have an assert on them that the device is AIE2.
- A few XLLVM ops are removed: we don't use broadcast or a few others,
so I've trimmed the set of ops we support down.
- This PR adds utils `isAie2` and `isAie2P` to AMDAIEUtils.h 


I have confirmed that a linalg matmul can compile all the way through
peano for AIE2P, but only with -O0. Next step after this is to fix
alignment issues in iree-amd-aie to get this to work for -On n>0:
https://github.com/Xilinx/llvm-aie/issues/315