Numerical test for strix compiled with peano   (#1077)
The main change here is an update to the pass that unsets the load and
store alignments. Now it unsets _all_ llvm.loads and llvm.stores (see
comments in pass). Without this, peano tries to scalarize an llvm.load
and gets into an infinite/quadratic loop. This PR adds an end-to-end
numerical test of peano on strix.

One unrelated change in AMDAIELowerToAIE.cpp : while experimenting with
adding alignment/noalias attributes to the function signature in
function outlining, noticed that the attributes were being dropped in
this pass.