[Test] Improve error message when no peano path (#1080)
The test added in 7fe13540b fails (cryptically) when `PEANO_INSTALL_DIR`
is not set (constructing std::string from nullptr). This PR improves the
diagnostics.

I guess we don't want peano to be in the minimal set of requirements to
run tests, so we should refactor this test to only run if
PEANO_INSTALL_DIR is present (or factorize aie2xlcbin somehow).