Support matmul 4d inputs on pack-peel-4-level-tiling (#1098)
This PR adds basic support for matmul with 4d inputs and output on
pack-peel-4-level-tiling.

In theory, the tiling strategy and pipeline should support different
layouts of matmul 4d operations. However, the order of the input dims
(both inner and outer dims) are crucial for the correct compilation and
results.

To ensure the correctness and for comparison purpose, this PR only adds
a test that corresponds to the standard matmul, which means the order of
the input dims for this operation corresponds to the L2 shapes of the
matmul op after the first level packing, i.e.,

```C += matmul4d(A,B) where A:MxKxM0xK0, B:NxKxK0xN0, C:NxMxM0xN0```
  
The test class and instance added in run.py is preliminary and for experimental purpose. Generalization of the test class will be addressed as follow-ups.

Runtime comparison:
On Phoenix runner `matmul_512_4096_512_bf16_f32 : 1141 us vs matmul4d_16_128_8_bf16_f32: 998 us`
On Strix runner `matmul_512_4096_512_i8_i32 : 806 us vs matmul4d_16_128_8_i8_i32: 491 us`