[ControlOverlay] Introduce `dma_placeholder` to preserve `connection` ops from Dead-Code Elimination (#1111)
Following the discussion in #1063, `connection` ops generated for the
control overlay do not initially have DMA users, as these are only added
later when the content of control packets is determined. However, since
`connection` ops are marked as pure in the IR, they can be wrongly
eliminated by CSE and Canonicalization if they have no users at that
stage.