[AMDAIESinkIntoCore] Generalize sinking for reuse with other regioned ops  (#1117)
This is part of the PR to vectorize linalg.fill:
https://github.com/nod-ai/iree-amd-aie/pull/1095

Basically one of the patterns introduced in
https://github.com/nod-ai/iree-amd-aie/pull/1095 means that in one of
the subsequent passes (lowering to llvm dialect) a cast operation is
introduced outside of an `aie.core`, which needs to be inside the aie.core
for core-to-standard to work. i.e. we need to sink an operation into an
aie.core. Before this PR, there is already a pass to sink operations
into `amdaie.core`. This PR refactors that pass so that it can be reused
to sink into `aie.core` (or any other regioned op).