[CtrlPkt] Add pipeline to dump control packet binaries to files (#1127)
This PR implements a pipeline in `XCLBinGen` for the generation and
conversion of control packets. The `aie.device` is converted into
`ctrlpkt_sequence.txt` and `ctrlpkt_instructions.txt`, which store the
contents and instructions for control packet DMAs, respectively. These
files will later be used by the driver to perform the reconfiguration.

Changes related to the driver and a new on-device test will be submitted
in follow-up PRs.