[CanonicalizeDoublyStridedOp] Add expansion of dimensions exceeding max sizes (#1144)
This PR adds a dimension expansion rewrite to the doubly-strided op
canonicalizer, which is effectively the reverse of the
`FoldDmaOpLinearDims` transformation. This fixes: #1137 (see test being
added).

Example:

```
offsets: [0], sizes: [8], strides: [1], maxSizes: [5]
```

is expanded into:

```
offsets: [0, 0], sizes: [2, 4], strides: [4, 1]
```

Here, all the sizes are smaller or equal to the max size. This helps
with enabling more DMA composition transformations, and this is needed
to avoid nested `npu.circular_dma_cpy_nd` ops after `DmaComposition`,
which is why it fixes #1137. Note that we should probably add some
checks that all `npu.circular_dma_cpy_nd` ops are combined and subsumed
correctly to help with debugging issues. I will address that in a
follow-up.