[CtrlPkt] Reconfiguration example for Matmul with constant bias (#1147)
This PR introduces a proof of concept for control packet reconfiguration
using a matmul operation with a constant bias. This example is chosen
because the differences between kernels are minimal, only the constant
bias varies, while other aspects, such as DMA connections and the rest
of the core code, remain unchanged.

This test is currently enabled on `xrt-lite` (linux) only. Changes on
`xrt` (windows) will be followed in another PR.