[DistributeL1Allocations] Add a fix for handling subviews in case of 1x1 core (#1189)
-- This commit adds a fix for handling memref.subviews in case of 1x1
core and for `pack-peel-4-level-tiling` which resolves a compilation
issue for `64 x 128 x 128` and thus an e2e test has been added for the
same.

**Previous consideration**
-- The pass `distribute-l1-allocations` considers only those
memref.subview ops as candidate for L1 alloc substitution which are
contained inside `scf.forall` with GPU thread IDs.

**Issue**
-- That assumption is correct for multiple core ops, but in case of 1x1
and for `pack-peel-4-level-tiling` the subview op need not follow that
and can be hoisted in the preceding `scf.forall` with NO GPU thread IDs.

**Fix**
-- Therefore this PR instead checks for memref.subview ops' users to be
within the innermost scf.forall ops as the latter is what constitutes
the computation of an individual core.
-- Therefore any L1 subview op with uses inside the innermost scf.forall
is a candidate for being replaced by a corresponding L1 alloc op.

Signed-off-by: Abhishek Varma <abhvarma@amd.com>

---------

Signed-off-by: Abhishek Varma <abhvarma@amd.com>