[CI] Fix operation count logging in test script (#1230)
This PR fixes the placement of logging for operation count, number of
rows, and number of columns in the test script. According to the parser
logic in
[performance_summarizer.py](https://github.com/nod-ai/iree-amd-aie/blob/5cfbf5cb0d1593dc3dc1a82d9122c36c9f81d7e1/build_tools/ci/cpu_comparison/performance_summarizer.py#L66-L68),
these values should be printed after the `Performance benchmark:` line.
Otherwise, it could cause `json_summary["tests"][-1]` to point to the
wrong entry.