[CI] Do not run benchmark tests twice on strix (#1267)
This PR updates the CI workflow to skip the Performance (benchmark)
tests during the `E2E comparison of AIE to llvm-cpu` step on Strix. The
corresponding numerical correctness tests will still run as expected.

No changes are needed for Phoenix tests, as the `--skip_tests` flag is
already set correctly for them.