Initial support of softmax uKernel (#1269)
This PR is a proof-of-concept and currently supports only the Chess and
Phoenix targets, uses the bfloat16 data type, and has only been tested
in a single-core configuration.

In the provided end-to-end example, the softmax operation has a shape of
128×32. The outer dimension (128) is tiled in two levels: first by 64,
then by 32. These numbers are currently randomly picked, and the
discovery of the optimal sizes is left for future work.

The uKernel implements an approximate version of softmax. As a result,
the current relative error tolerance is set to a relatively high value
of 2e-2. Within the uKernel, computation along the outer dimension is
performed sequentially, while the inner dimension is vectorized with a
chunk size of 16.