[KernelDispatch] Generalize softmax-copy pipeline to support elementwise or reduction ops (#1297)
- The copy based pipeline can be reused for both elementwise and
reduction ops while the root configs for these ops can be different.
This PR adds basic tiling supports to these ops and renames
`softmax-copy` pipeline as `general-copy` pipeline.

- For elementwise ops, we currently only support 1D generic input,
because `CollapseDimensionsPass` during dispatch creation will collapse
all higher dimensions to one dimension. We added this support because it
is helpful for developing vectorization for some elementwise ops,
however, in real application, we would expect the elementwise op as a
producer/consumer and fused with other compute ops in a single dispatch.
So a separate pipeline for elementwise ops is not very useful.

- The elementwise op example (as shown below) currently fails at
`AMDAIEDistributeCoresAndObjectFifos`.

```
func.func @truncf(%arg0: tensor<128xf32>) -> tensor<128xbf16> {
  %0 = tensor.empty() : tensor<128xbf16>
  %1 = linalg.generic {indexing_maps = [affine_map<(d0) -> (d0)>, affine_map<(d0) -> (d0)>], iterator_types = ["parallel"]} ins(%arg0 : tensor<128xf32>) outs(%0 : tensor<128xbf16>) {
  ^bb0(%in: f32, %out: bf16):
    %1 = arith.truncf %in : f32 to bf16
    linalg.yield %1 : bf16
  } -> tensor<128xbf16>
  return %1 : tensor<128xbf16>
}
```

- The reduction op example works e2e without vectorization, and the next
step is to support reduction dimension tiling and vectorization.