Use pip install in peano download script (#1303)
Previously, we used `pip download` and then extracted the .whl manually.
However, this approach fails when there are multiple versions of
llvm_aie wheels present in the same directory, as the script cannot
identify the correct .whl to unzip, and will raise an error.

To make the process more simple and robust, this PR now uses `pip
install` directly with the `--target` option to install/overwrite
llvm_aie to a specified directory.

One drawback of using `--target` is that it disables pip’s caching
mechanism, resulting in the .whl file (~140MB) being downloaded each
time the script runs, even if the version hasn’t changed. That said,
this doesn't appear to significantly impact CI runtimes.