[SplitLogicalObjFifo] Avoid over-splitting if the op shape is small  (#1310)
The split factor was previously determined based on the number of
columns available on the target device. However, when the operation
shape is too small to utilize all available columns, this led to
over-splitting.

This PR refines the logic by also inspecting the CoreOps present in the
module. If cores are already assigned and occupy only a subset of the
columns, the split factor is adjusted accordingly based on the actual
number of columns in use.