[uKernel] Determine chess zero-filled vectorization size based on data type (#1321)
Previously, the vectorization size r was set to N or N/2, which could
cause numerical issues for large values of N due to exceeding the
bitwidth limits of the store unit. This PR updates the logic to directly
determine r based on the data type instead. Some unused template
function instantiations are also removed.